--- src/common/utils/buildCommonTranslatedDictionary.js
@@ -1,9 +1,0 @@
-import invert from 'lodash/invert';
-import { FormattedMessage } from 'react-intl';
-
-export const buildCommonTranslatedDictionary = (dictionary, baseTranslationKey) => {
-  return Object.fromEntries(
-    Object.entries(invert(dictionary))
-      .map(([orderType, key]) => [orderType, <FormattedMessage key={key} id={`${baseTranslationKey}${key}`} />]),
-  );
-};
