+++ src/common/utils/index.js
@@ -1,4 +1,3 @@
-export { buildCommonTranslatedDictionary } from './buildCommonTranslatedDictionary';
 export { buildRelationModifierQuery } from './buildRelationModifierQuery';
 export * from './checkRelatedHoldings';
 export * from './createClearFilterHandler';
